+++ framework-system/framework-management/src/main/java/uk/gov/justice/services/management/shuttering/handler/ShutteringSystemCommandHandler.java
@@ -1,12 +1,12 @@
 package uk.gov.justice.services.management.shuttering.handler;
 
-import static uk.gov.justice.services.jmx.api.command.ShutterSystemCommand.SHUTTER_APPLICATION;
-import static uk.gov.justice.services.jmx.api.command.UnshutterSystemCommand.UNSHUTTER_APPLICATION;
+import static uk.gov.justice.services.jmx.api.command.ShutterSystemCommand.SHUTTER;
+import static uk.gov.justice.services.jmx.api.command.UnshutterSystemCommand.UNSHUTTER;
 
 import uk.gov.justice.services.common.util.UtcClock;
-import uk.gov.justice.services.jmx.command.HandlesSystemCommand;
 import uk.gov.justice.services.jmx.api.command.ShutterSystemCommand;
 import uk.gov.justice.services.jmx.api.command.UnshutterSystemCommand;
+import uk.gov.justice.services.jmx.command.HandlesSystemCommand;
 import uk.gov.justice.services.management.shuttering.events.ShutteringRequestedEvent;
 import uk.gov.justice.services.management.shuttering.events.UnshutteringRequestedEvent;
 
@@ -24,12 +24,12 @@ public class ShutteringSystemCommandHandler {
     @Inject
     private Event<UnshutteringRequestedEvent> unshutteringRequestedEventFirer;
 
-    @HandlesSystemCommand(SHUTTER_APPLICATION)
+    @HandlesSystemCommand(SHUTTER)
     public void onShutterRequested(final ShutterSystemCommand shutterSystemCommand) {
         shutteringRequestedEventFirer.fire(new ShutteringRequestedEvent(shutterSystemCommand, clock.now()));
     }
 
-    @HandlesSystemCommand(UNSHUTTER_APPLICATION)
+    @HandlesSystemCommand(UNSHUTTER)
     public void onUnshutterRequested(final UnshutterSystemCommand unshutterSystemCommand) {
         unshutteringRequestedEventFirer.fire(new UnshutteringRequestedEvent(unshutterSystemCommand, clock.now()));
     }