+++ jmx/jmx-api/src/main/java/uk/gov/justice/services/jmx/api/command/UnshutterSystemCommand.java
@@ -2,10 +2,10 @@ package uk.gov.justice.services.jmx.api.command;
 
 public class UnshutterSystemCommand extends BaseSystemCommand {
 
-    public static final String UNSHUTTER_APPLICATION = "UNSHUTTER_APPLICATION";
+    public static final String UNSHUTTER = "UNSHUTTER";
     private static final String DESCRIPTION = "Unshutters the application.";
 
     public UnshutterSystemCommand() {
-        super(UNSHUTTER_APPLICATION, DESCRIPTION);
+        super(UNSHUTTER, DESCRIPTION);
     }
 }